Reduced the number of properties in the UserSchema. Only takes auth0id, username, and friends array now.
Edited app.js to specify if an error has been caused by an expired token.

Deleted some user routes and altered the /user route to be /user/register.

The user_create_post controller function decodes the token and registers the user to MongoDB, if the user does not already exist.