Refactor makeAuthenticatedRequest to include an options object to hold the user.
Sends the options.user object in the headers. Note: can't send in data unless the request is a put or post method.

makeAuthenticatedRequest call now includes a data object and an options object in applicable functions.

isAuthenticated and user objects are obtained from useAuth0 hook in applicable functions.

When getting conversations and messages, checks for isAuthenticated. isAuthenticated is a dependency in useEffect hook for rendering conversations and messages in ConvoPage and ConvosPage.